Refactor package structure and file names in internal/command_test directory to internal/test_cases. This change improves code organization and follows best practices for package naming. The affected files are command_test_case.go, receive_value_test_case.go, and send_replication_handshake_test_case.go. Additionally, update import statements in test_echo.go, test_expiry.go, test_get_set.go, test_ping_pong.go, and test_repl_master_cmd_prop.go to reflect the new package structure. These changes enhance code readability and maintainability.
