+++ internal/test_cases/send_replication_handshake_test_case.go
@@ -1,4 +1,4 @@
-package command_test
+package test_cases
 
 import (
 	"bytes"
